Move csvUploader method to StorageHandler class
